fix: config type bugs, null json crash, transfer cancel method
- Cast env('ASAAS_TIMEOUT') to int in config to prevent TypeError
- Validate api_key in ServiceProvider with descriptive RuntimeException
- Guard against null json in toResult() with ?? [] for empty 2xx responses
- Change TransferResource::cancel() from DELETE to POST (consistent with
  all other cancel endpoints: PixTransaction, Invoice, BillPayment)
- Add .env to .gitignore to prevent accidental API key commits

+++ config/asaas.php
@@ -5,5 +5,5 @@ declare(strict_types=1);
 return [
     'api_key' => env('ASAAS_API_KEY'),
     'environment' => env('ASAAS_ENVIRONMENT', 'sandbox'),
-    'timeout' => env('ASAAS_TIMEOUT', 30),
+    'timeout' => (int) env('ASAAS_TIMEOUT', 30),
 ];
